server: split size_spec off from store_spec
This commit splits these types into separate files which helps with code
cleanup and re-org.

Epic: none

Release note: None
